"[Add/#127] social login refactoring"
