+++ app/src/main/java/com/example/umc/Signin/KakaoLoginManager.kt
@@ -1,4 +1,4 @@
-package com.example.umc.UserApi
+package com.example.umc.Signin
 
 import android.content.Context
 import android.provider.Settings
