+++ app/src/main/res/layout/fragment_animation.xml
@@ -79,10 +79,11 @@
         android:id="@+id/textView20"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/pretendard_bold"
         android:layout_marginTop="28sp"
+        android:fontFamily="@font/pretendard_bold"
         android:text="님의"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.529"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/imageView10" />
